K closest points
use a heap to achieve the target. First calculate all the distances, then add first k elements in the heap, for rest of the heap, if the distance is larger than the top element in the heap, do nothing, if it's smaller, than pop the current top element and add the new element, this way to make sure the heap will store the k smallest elements.